孙林添加框架接口以及随机生成士兵
1. 添加了SharedKey中新的借口（等待王楠实现）
2. 修改了target中的bigInteger为int，并且顺带修改了sceneCreator中的bigInteger为int
3. xmlReader中支持随机生成士兵

+++ src/thu/infosecurity/simulate/controller/SceneCreator.java
@@ -82,7 +82,7 @@ public class SceneCreator {
         /* 生成装备箱列表 */
         ArrayList<Target> targetList = new ArrayList<>();
         for(int i = 0; i < targetNum; i++){
-            Target tgt = new Target("TackleBox1",new Point(Utils.generateRandom(0,1000), Utils.generateRandom(0,800)),"S","G",true,3, new BigInteger("1234"));
+            Target tgt = new Target("TackleBox1",new Point(Utils.generateRandom(0,1000), Utils.generateRandom(0,800)),"S","G",true,3, new Integer("1234"));
             targetList.add(tgt);
         }
